Move Supabase utility code up into common
